+++ changelog.md
@@ -1,4 +1,5 @@
 ## unreleased
+- CO-2686 fix missing customer website
 
 ## 2.0.1 _2024-07-05_
 - Fixed type errors