+++ changelog.md
@@ -4,6 +4,7 @@
 - CO-2786 - fix missing tab name
 - CO-2764 - Prestashop 8.2 compatibility
 - CO-2773 - add gpsr metadata to product description
+- CO-2853 - fix language iso type
 
 ## 2.0.1 _2024-07-05_
 - Fixed type errors